requirements-dev.txt
-r requirements.txt
bandit >= 1.6.2
black >= 20.8b1
coverage >= 5.3
flake8 >= 3.8.4
pylint >= 2.6.0
pytest >= 6.1.2
pytest-cov >= 2.10.1
discord.py == 1.7.3